kqueue: Support new `struct kevent` in FreeBSD 12.x
In FreeBSD 12.x, time-related fields were changed to 64-bit, and
re-ordered. That part is only a concern for libc. However, the structure
also grew a new member, `ext`, and we need to initialize it.

A 0-length `ext` array was also added to The definition of `struct
kevent` for FreeBSD 11.x in the libc to ease the use of this structure.
We use the new `libc::KEVENT_EXT_ZEROED` constant which helps initialize
that member, regardless of the version of FreeBSD.